Create GUI window with a basic menu bar
Currently, the light theme is always used, but the code can be easily
changed to use the dark theme instead. In the future, this should be
loaded from some kind of user preferences file, but I don't know what
file format we'll be using for that yet.

A few Java system properties are set on startup to make the application
work better on macOS. I don't have a Mac to test these, but they should
make the application use the macOS menu bar and set the window title bar
appearance.

The application's name (used in the window title and About menu item)
can also be easily changed because I'm not sure what it should be.