Install the FlatLaf inspector
This is similar to the Inspect Element feature of web browsers, but for
Java Swing instead. When enabled, it can be activated by pressing
Ctrl+Alt+Shift+X. Additionally, the UI default inspector, which shows
UI variables used by Swing, can be opened with Ctrl+Alt+Shift+Y.

The inspector is disabled by default. To enable the inspector, add
-Dcs2212.mapmaker.enableInspector to the JVM options used to launch the
application (in IntelliJ, this is in your run configuration settings).