Implement core of help viewer
The layout of the help viewer consists of a tree view of help pages on
the left and the contents of one page to the right. This is fairly
standard for help and documentation viewers AFAIK.

Currently, the help viewer will load the list of pages from a JSON file
and will load the contents of the help home page. Navigation has not
been implemented yet.