Make core classes immutable
The biggest benefit of this change is that it allows us to use Java
records to implement these classes instead of having to write lots of
getter/setter soup. This is especially important as previously each
getter and setter would have required an individual Javadoc comment (per
the project specification).

Updating information stored in these classes should be accomplished by
creating a new object with the same values except for what you want to
change.