Create new version of map screen
This commit removes the previous map screen that essentially looked like
widget soup (sorry Arjun) and adds a new map screen that more closely
follows the mockup. Only the toolbar at the top and the sidebar on the
right are implemented right now, and none of the functionality is here.