Add new Database class
This commit introduces the third version of the Database class. This
time, it is responsible for file I/O (loading and saving the database).

This is accomplished by holding a FileChannel open, then reading and
writing to it as required. On Windows, holding the file open will
prevent other applications from modifying the file while it is open.