Changed LoginWindow.java into MainWindow.java
Changed named of LoginWindow to MainWindow as it makes more sense, since
the LoginWindow was just acting as a main window. Moved it outside the
log in directory as it made since it is no longer part of the log in
stuff. Deleted previous MainWindow since it was not needed.