Switch to a simpler save/load strategy
Instead of holding the file open for the entire lifetime of the
application, the file will be opened on-demand.

Since opening and closing files is a potentially expensive operation,
this strategy might have worse performance, but it reduces the
complexity of the code and removes the requirement to close the database
when the application exits.

I'm still not entirely sure which implementation is better. However, the
code for the old implementation still exists; just revert this commit to
bring it back.