Implement developer log in & sign in error
Made placeHolder accounts, one that is a developer and one that is a user. You can enter either "developer" or "regular user" on the sign-in page without a password to select which mode.

Currently only the title will change depending on who is logged in.

Created sign in error in the case the username is not recognized.

Placeholder accounts will get replaced by database functions once Database has been merged.