Add data transfer for accounts with database
Program can now access the database to see whether the user has entered
a valid username. Currently, we have 2 accounts, one for developer and
one for a regular user.

Note that there is no password for either of them and we have not added
functionality to create accounts just yet.

Database implemenation with the building will be added once the database
has been updated with the buildings and POIS.