Various minor improvements to code
Removed weird Nullable and checking for null in the Map panel; we check
for null in Main panel now.

Made database a simple in the MainWindow; intellij gave me a warning
about this. If we need to make it back into a field it'll be easy

Removed the LogedinAccount field in the Map panel; we just use the
argument that is given for now.

Removed the database addition in Create Account; we can add it back
later if we need to.