Begin implement search function
Added a search bar that shows when you press enter on the search bar. It
will disappear if you select an element on the list and then select
something else

Currently, it displays only the thing you entered the search bar, will
add proper searching once floor implementation has been completed.

As well, the visual aspect of the search list is buggy, and needs work.