Make Jackson output JSON that meets style rules
This allows the application to be used to edit the default database
without worrying about reformatting the JSON file later (even if that
would be done automatically).

The changes made include sorting properties alphabetically, adding
two-space indentation, and ensuring there is a newline at the end of the
file.

This commit also replaces the JSON file with one generated by Jackson,
which added "favorite" fields to everything and regenerated floor IDs.