Implement building the final distribution ZIP
This distribution ZIP includes the executable JAR file for the
application, all runtime dependencies, the built-in database, and any
README, LICENSE, or NOTICE files in the project root.

You can build the ZIP using the `mvn clean package` command and the
resulting ZIP will be found in the dist folder.