Make favorite POIs specific to each account
This commit makes it so the favorites data is stored as `Set<Account>
favoriteOf` instead of `boolean favorite`. This requires extensive
changes across the codebase to make the old code work with the new data
format.

One thing to note is that accounts now have an "@id" property like
floors do. This requires old database JSON files to be converted, which
right now must be done manually.